Moved AudioEmitter to correct package
